+++ CHANGELOG.md
@@ -173,3 +173,4 @@
 - 2023-11-16: Add follow up, sender, users, deal, +3 types
 - 2023-11-17: Wire discover, find email, email opener, lib, +6
 - 2023-11-17: Shutdown find email, icp, configuration, Makefile, +6
+- 2023-11-18: Add tz country, lib, users, descend, +6 types